+++ tests/PhpWord/Tests/Writer/Word2007/BaseTest.php
@@ -159,6 +159,11 @@ class BaseTest extends \PHPUnit_Framework_TestCase
             $section->addImage(__DIR__ . "/../../_files/images/earth.jpg", $styles);
         }
 
+        $archiveFile = realpath(__DIR__ . '/../../_files/documents/reader.docx');
+        $imageFile = 'word/media/image1.jpeg';
+        $source = 'zip://' . $archiveFile . '#' . $imageFile;
+        $section->addImage($source);
+
         $doc = TestHelperDOCX::getDocument($phpWord);
 
         // behind